[zep fromtree] refactor(rcar): rename console_rcar_ to console_renesas_ prefix for Renesas platform
Rename console_rcar_ to console_renesas_ prefix for SCIF-based console
driver to make it reusable by other Renesas platforms.

Due to the above renaming, function console_renesas_register is duplicated
in both scif.h and console.h, so it should be removed from scif.h

(cherry picked from commit 989362581c6bd1b894ad943204ec97cbe81d12ee)
