+++ drivers/renesas/common/console/rcar_console.S
@@ -9,12 +9,12 @@
 #include <console_macros.S>
 #include <drivers/renesas/rcar/console/console.h>
 
-	.globl	console_rcar_register
+	.globl	console_renesas_register
 
-	.extern	rcar_log_init
+	.extern	renesas_log_init
 
 	/* -----------------------------------------------
-	 * int console_rcar_register(
+	 * int console_renesas_register(
 	 *      uintptr_t base, uint32_t clk, uint32_t baud,
 	 *      console_t *console)
 	 * Function to initialize and register a new rcar
@@ -28,19 +28,19 @@
 	 * Clobber list : x0, x1, x2, x6, x7, x14
 	 * -----------------------------------------------
 	 */
-func console_rcar_register
+func console_renesas_register
 	mov	x7, x30
 	mov	x6, x3
 	cbz	x6, register_fail
 	str	x0, [x6, #CONSOLE_T_BASE]
 
-	bl	rcar_log_init
+	bl	renesas_log_init
 	cbz	x0, register_fail
 
 	mov	x0, x6
 	mov	x30, x7
-	finish_console_register rcar, putc=1, getc=0, flush=1
+	finish_console_register renesas, putc=1, getc=0, flush=1
 
 register_fail:
 	ret	x7
-endfunc console_rcar_register
+endfunc console_renesas_register
