+++ drivers/renesas/common/scif/scif.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021-2025, Renesas Electronics Corporation. All rights reserved.
+ * Copyright (c) 2021-2026, Renesas Electronics Corporation. All rights reserved.
  *
  * SPDX-License-Identifier: BSD-3-Clause
  */
@@ -7,12 +7,10 @@
 #ifndef SCIF_H
 #define SCIF_H
 
-int console_rcar_register(uintptr_t base, uint32_t clk, uint32_t baud,
-			  console_t *pconsole);
 void scif_console_set_regs(uint32_t fsr, uint32_t tdr);
-int console_rcar_init(uintptr_t base_addr, uint32_t uart_clk,
+int console_renesas_init(uintptr_t base_addr, uint32_t uart_clk,
 		      uint32_t baud_rate);
-int console_rcar_putc(int c, console_t *pconsole);
-int console_rcar_flush(console_t *pconsole);
+int console_renesas_putc(int c, console_t *pconsole);
+int console_renesas_flush(console_t *pconsole);
 
 #endif /* SCIF_H */
