+++ drivers/renesas/common/scif/scif_helpers.S
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2025, Renesas Electronics Corporation. All rights reserved.
+ * Copyright (c) 2025-2026, Renesas Electronics Corporation. All rights reserved.
  *
  * SPDX-License-Identifier: BSD-3-Clause
  */
@@ -8,15 +8,15 @@
 #include <asm_macros.S>
 #include <console_macros.S>
 
-	.globl	console_rcar_register
-	.extern	console_rcar_init
+	.globl	console_renesas_register
+	.extern	console_renesas_init
 
 	/*
 	 * -----------------------------------------------
-	 * int console_rcar_register(
+	 * int console_renesas_register(
 	 *      uintptr_t base, uint32_t clk, uint32_t baud,
 	 *      console_t *console)
-	 * Function to initialize and register a new rcar
+	 * Function to initialize and register a new Renesas
 	 * console. Storage passed in for the console struct
 	 * *must* be persistent (i.e. not from the stack).
 	 * In: x0 - UART register base address
@@ -27,18 +27,18 @@
 	 * Clobber list : x0, x1, x2, x6, x7, x14
 	 * -----------------------------------------------
 	 */
-func console_rcar_register
+func console_renesas_register
 	mov	x7, x30
 	mov	x6, x3
 	cbz	x6, register_fail
 	str	x0, [x6, #CONSOLE_T_BASE]
 
-	bl	console_rcar_init
+	bl	console_renesas_init
 
 	mov	x0, x6
 	mov	x30, x7
-	finish_console_register rcar, putc=1, getc=0, flush=1
+	finish_console_register renesas, putc=1, getc=0, flush=1
 
 register_fail:
 	ret	x7
-endfunc console_rcar_register
+endfunc console_renesas_register
