+++ drivers/renesas/rcar_gen4/scif/scif.c
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021-2025, Renesas Electronics Corporation. All rights reserved.
+ * Copyright (c) 2021-2026, Renesas Electronics Corporation. All rights reserved.
  *
  * SPDX-License-Identifier: BSD-3-Clause
  */
@@ -60,7 +60,7 @@
 #define MODEMR_HSCIF_DLMODE_1843200	2U
 #define MODEMR_HSCIF_DLMODE_3000000	3U
 
-int console_rcar_init(uintptr_t base_addr, uint32_t uart_clk,
+int console_renesas_init(uintptr_t base_addr, uint32_t uart_clk,
 		      uint32_t baud_rate)
 {
 	uint32_t modemr, mstpcr, mstpsr, mstpbit;
