+++ drivers/renesas/rcar_gen5/scif/scif.c
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2025, Renesas Electronics Corporation. All rights reserved.
+ * Copyright (c) 2025-2026, Renesas Electronics Corporation. All rights reserved.
  *
  * SPDX-License-Identifier: BSD-3-Clause
  */
@@ -18,7 +18,7 @@
 #define HSCIF_HSFTDR	(HSCIF_BASE + 0x000CU) /*  8 Transmit FIFO data register */
 #define HSCIF_HSFSR	(HSCIF_BASE + 0x0010U) /* 16 Serial status register */
 
-int console_rcar_init(uintptr_t base_addr, uint32_t uart_clk,
+int console_renesas_init(uintptr_t base_addr, uint32_t uart_clk,
 		      uint32_t baud_rate)
 {
 	scif_console_set_regs(HSCIF_HSFSR, HSCIF_HSFTDR);
