+++ plat/renesas/common/aarch64/plat_helpers.S
@@ -1,6 +1,6 @@
 /*
  * Copyright (c) 2013-2023, Arm Limited and Contributors. All rights reserved.
- * Copyright (c) 2015-2025, Renesas Electronics Corporation. All rights reserved.
+ * Copyright (c) 2015-2026, Renesas Electronics Corporation. All rights reserved.
  *
  * SPDX-License-Identifier: BSD-3-Clause
  */
@@ -28,9 +28,9 @@
 	.globl	plat_my_core_pos
 	.extern	rcar_log_init
 
-	.extern console_rcar_init
-	.extern console_rcar_putc
-	.extern console_rcar_flush
+	.extern console_renesas_init
+	.extern console_renesas_putc
+	.extern console_renesas_flush
 
 #if IMAGE_BL2
 	#define	INT_ID_MASK	(0x3ff)
@@ -270,7 +270,7 @@ func plat_crash_console_init
 	mov	sp, x2
 	str	x1, [sp, #-16]!
 	str	x30, [sp, #-16]!
-	bl	console_rcar_init
+	bl	console_renesas_init
 	ldr	x30, [sp], #16
 	ldr	x1, [sp], #16
 	mov	sp, x1
@@ -294,7 +294,7 @@ func plat_crash_console_putc
 	str	x5, [sp, #-16]!
 	str	x6, [sp, #-16]!
 	str	x7, [sp, #-16]!
-	bl	console_rcar_putc
+	bl	console_renesas_putc
 	ldr	x7, [sp], #16
 	ldr	x6, [sp], #16
 	ldr	x5, [sp], #16
@@ -311,7 +311,7 @@ endfunc plat_crash_console_putc
 	 * ---------------------------------------------
 	 */
 func plat_crash_console_flush
-	b	console_rcar_flush
+	b	console_renesas_flush
 endfunc plat_crash_console_flush
 
 	/* --------------------------------------------------------------------
