+++ plat/renesas/rcar_gen4/aarch64/plat_helpers.S
@@ -1,6 +1,6 @@
 /*
  * Copyright (c) 2013-2025, ARM Limited and Contributors. All rights reserved.
- * Copyright (c) 2015-2025, Renesas Electronics Corporation. All rights reserved.
+ * Copyright (c) 2015-2026, Renesas Electronics Corporation. All rights reserved.
  *
  * SPDX-License-Identifier: BSD-3-Clause
  */
@@ -27,9 +27,9 @@
 	.globl	plat_renesas_calc_core_pos
 	.globl	bl31_plat_enable_mmu
 
-	.extern console_rcar_init
-	.extern console_rcar_putc
-	.extern console_rcar_flush
+	.extern console_renesas_init
+	.extern console_renesas_putc
+	.extern console_renesas_flush
 	.extern rcar_pwrc_code_copy_to_system_ram
 
 	/* -----------------------------------------------------
@@ -176,7 +176,7 @@ func plat_crash_console_init
 	mov	sp, x2
 	str	x1, [sp, #-16]!
 	str	x30, [sp, #-16]!
-	bl	console_rcar_init
+	bl	console_renesas_init
 	ldr	x30, [sp], #16
 	ldr	x1, [sp], #16
 	mov	sp, x1
@@ -199,7 +199,7 @@ func plat_crash_console_putc
 	str	x5, [sp, #-16]!
 	str	x6, [sp, #-16]!
 	str	x7, [sp, #-16]!
-	bl	console_rcar_putc
+	bl	console_renesas_putc
 	ldr	x7, [sp], #16
 	ldr	x6, [sp], #16
 	ldr	x5, [sp], #16
@@ -216,7 +216,7 @@ endfunc plat_crash_console_putc
 	 * ---------------------------------------------
 	 */
 func plat_crash_console_flush
-	b	console_rcar_flush
+	b	console_renesas_flush
 endfunc plat_crash_console_flush
 
 	/* ---------------------------------------------
